CAUSEWAY-3815: working on encapsulation
