+++ viewers/wicket/ui/src/main/java/org/apache/causeway/viewer/wicket/ui/components/actions/ActionParametersPanelFactory.java
@@ -44,7 +44,7 @@ public class ActionParametersPanelFactory extends ComponentFactoryAbstract {
     @Override
     public Component createComponent(final String id, final IModel<?> model) {
         final ActionModel actionModel = (ActionModel) model;
-        return new ActionParametersPanel(id, actionModel);
+        return new ActionParametersPanel(id, actionModel, false);
     }
 
 }
